Update to reflect documentation
Documentation shows the use for payments as passing a team_reference in as an integer. Passing in as an integer fails with "TypeError: no implicit conversion of Fixnum into String". Changing this post to use string interpolation fixes the issues and now reflects the proper use per the API documentation: https://developers.upwork.com/?lang=ruby#payments_make-custom-payment